Ajoute une description pour la recherche APE
