+++ site/source/locales/ui-fr.yaml
@@ -1304,6 +1304,19 @@ pages:
     à&nbsp;:<1></1>Défenseur des droits<3></3>Libre réponse
     71120<5></5>75342&nbsp;Paris CEDEX 07</2></21><22>Mis à jour le
     29/01/2021</22>
+  assistant:
+    recherche-code-ape:
+      seo explanation: "<0>Comment fonctionne la recherche de code APE ?</0><1>Pour
+        retrouver le code APE de votre entreprise, saisissez un ou deux
+        mots-clés qui décrivent l'activité exercée. Par exemple : « restauration
+        » ou « coach sportif » La recherche utilise les bases de l'Insee et du
+        Guichet Unique pour trouver les activités qui correspondent. Par
+        ailleurs, les résultats les plus populaires (avec le plus d'entreprise
+        crées) sont affichés en premier.</1><2>Si vous ne retrouvez pas le code
+        APE correspondant à votre activité, vous pouvez
+        nous le signaler avec le bouton « Je ne trouve pas mon activité ». Cela
+        nous permettra d'améliorer la recherche en ajoutant des mots clés au fur
+        et à mesure de vos retours.</2>"
   assistants:
     pour-mon-entreprise:
       avertissement-entreprise-non-traitée: <0>Il n'existe pas encore de simulateur de
@@ -1331,11 +1344,12 @@ pages:
         description: Simulateurs et assistants adaptés à votre entreprise
       title: Gérer mon activité
     recherche-code-ape:
-      description: Cet assistant vous permet de trouver rapidement le code APE
+      description: <0>Cet assistant vous permet de trouver rapidement le code APE
         (activité principale exercée) qui correspond à votre activité. Ce code
         est attribué par l'INSEE à chaque entreprise en France afin de les
-        catégoriser. Retrouvez plus d'informations sur le code APE sur
-        <2>entreprendre.service-public.fr</2> et <6>economie.gouv.fr</6>.
+        catégoriser.</0><1><0>Retrouvez plus d'informations sur le code APE sur
+        <2>entreprendre.service-public.fr</2> et
+        <6>economie.gouv.fr</6>.</0></1>
       meta:
         description: Assistant pour trouver le code APE qui correspond à votre activité.
         title: Recherche de code APE
